GRIDEDIT-757 Moved mesh conversion class and concepts to separate file.
